separate reader and writer test files
